feat(dashboard): Expose bulk actions per row in data tables
